add cast to u_long before cast to caddr_t make these compile on sparc64

+++ sys/dev/sbus/cgthree_sbus.c
@@ -1,4 +1,4 @@
-/*	$NetBSD: cgthree_sbus.c,v 1.3 2001/09/26 20:53:14 eeh Exp $ */
+/*	$NetBSD: cgthree_sbus.c,v 1.4 2001/10/05 14:25:15 pooka Exp $ */
 
 /*-
  * Copyright (c) 1998 The NetBSD Foundation, Inc.
@@ -191,7 +191,7 @@ cgthreeattach_sbus(parent, self, args)
 	name = PROM_getpropstring(node, "model");
 
 	if (sa->sa_npromvaddrs != 0)
-		fb->fb_pixels = (caddr_t)sa->sa_promvaddrs[0];
+		fb->fb_pixels = (caddr_t)(u_long)sa->sa_promvaddrs[0];
 	if (isconsole && fb->fb_pixels == NULL) {
 		int ramsize = fb->fb_type.fb_height * fb->fb_linebytes;
 		if (sbus_bus_map(sa->sa_bustag, sa->sa_slot,
